worked on #364 (Improve GP evaluation performance)  * removed list of Instr  * changes that didn't affect performance directly: reduced size of Instr class, added 'constant-folding' and pre-calculation of skip-lengths
git-svn-id: https://src.heuristiclab.com/svn/core/trunk@767 2abd9481-f8db-48e9-bd25-06bc13291c1b

+++ sources/HeuristicLab.GP.StructureIdentification/SymbolTable.cs
@@ -28,37 +28,37 @@ using System.Xml;
 
 namespace HeuristicLab.GP.StructureIdentification {
   class EvaluatorSymbolTable : StorableBase {
-    public const int ADDITION = 1;
-    public const int AND = 2;
-    public const int AVERAGE = 3;
-    public const int CONSTANT = 4;
-    public const int COSINUS = 5;
-    public const int DIFFERENTIAL = 25;
-    public const int DIVISION = 6;
-    public const int EQU = 7;
-    public const int EXP = 8;
-    public const int GT = 9;
-    public const int IFTE = 10;
-    public const int LT = 11;
-    public const int LOG = 12;
-    public const int MULTIPLICATION = 13;
-    public const int NOT = 14;
-    public const int OR = 15;
-    public const int POWER = 16;
-    public const int SIGNUM = 17;
-    public const int SINUS = 18;
-    public const int SQRT = 19;
-    public const int SUBTRACTION = 20;
-    public const int TANGENS = 21;
-    public const int VARIABLE = 22;
-    public const int XOR = 23;
-    public const int UNKNOWN = 24;
+    public const byte ADDITION = 1;
+    public const byte AND = 2;
+    public const byte AVERAGE = 3;
+    public const byte CONSTANT = 4;
+    public const byte COSINUS = 5;
+    public const byte DIFFERENTIAL = 25;
+    public const byte DIVISION = 6;
+    public const byte EQU = 7;
+    public const byte EXP = 8;
+    public const byte GT = 9;
+    public const byte IFTE = 10;
+    public const byte LT = 11;
+    public const byte LOG = 12;
+    public const byte MULTIPLICATION = 13;
+    public const byte NOT = 14;
+    public const byte OR = 15;
+    public const byte POWER = 16;
+    public const byte SIGNUM = 17;
+    public const byte SINUS = 18;
+    public const byte SQRT = 19;
+    public const byte SUBTRACTION = 20;
+    public const byte TANGENS = 21;
+    public const byte VARIABLE = 22;
+    public const byte XOR = 23;
+    public const byte UNKNOWN = 24;
 
-    private static Dictionary<Type, int> staticTypes = new Dictionary<Type, int>();
+    private static Dictionary<Type, byte> staticTypes = new Dictionary<Type, byte>();
 
     // needs to be public for persistence mechanism (Activator.CreateInstance needs empty constructor)
     static EvaluatorSymbolTable() {
-      staticTypes = new Dictionary<Type, int>();
+      staticTypes = new Dictionary<Type, byte>();
       staticTypes[typeof(Addition)] = ADDITION;
       staticTypes[typeof(And)] = AND;
       staticTypes[typeof(Average)] = AVERAGE;
@@ -85,7 +85,7 @@ namespace HeuristicLab.GP.StructureIdentification {
       staticTypes[typeof(Differential)] = DIFFERENTIAL;
     }
 
-    internal static int MapFunction(IFunction function) {
+    internal static byte MapFunction(IFunction function) {
       if(staticTypes.ContainsKey(function.GetType())) return staticTypes[function.GetType()];
       else return UNKNOWN;
     }
